requirements.txt
# PDF processing
PyMuPDF>=1.24.0

# Gemini (embedding + generation)
google-genai>=1.0.0

# Vector DB (no Docker needed, pure Python)
chromadb>=0.5.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0